feat(feedback): убрать плавающий виджет обратной связи, сделать отдельным экраном
Раньше форма обратной связи открывалась плавающей кнопкой "Напишите нам"
в правом нижнем углу поверх каждой страницы сайта (FeedbackWidgetButton +
FeedbackWidgetProvider + модалка-портал). Теперь это отдельный экран
/feedback, доступный только по ссылке "Обратная связь" в футере (раздел
"О проекте") — без плавающего триггера на каждой странице.

Форма перенесена как есть (валидация, текст успеха, обработка ошибок),
виджет и провайдер удалены целиком как больше не нужные.
