+++ public/locales/en/translation.json
@@ -104,7 +104,8 @@
                     "privacy-policy": "Privacy policy",
                     "news": "News",
                     "about-goodsurfing": "About Goodsurfing",
-                    "find-job": "Paid Work"
+                    "find-job": "Paid Work",
+                    "feedback": "Feedback"
                 },
                 "for-organizers": {
                     "title": "For organizers",
