+++ public/locales/es/translation.json
@@ -104,7 +104,8 @@
                     "privacy-policy": "Política de Privacidad",
                     "news": "Noticias",
                     "about-goodsurfing": "Sobre Goodsurfing",
-                    "find-job": "Trabajo remunerado"
+                    "find-job": "Trabajo remunerado",
+                    "feedback": "Comentarios"
                 },
                 "for-organizers": {
                     "title": "Para los organizadores",
