+++ public/locales/ru/translation.json
@@ -104,7 +104,8 @@
                     "privacy-policy": "Политика конфиденциальности",
                     "news": "Новости",
                     "about-goodsurfing": "О Гудсёрфинге",
-                    "find-job": "Оплачиваемая работа"
+                    "find-job": "Оплачиваемая работа",
+                    "feedback": "Обратная связь"
                 },
                 "for-organizers": {
                     "title": "Для организаторов",
