--- src/app/providers/FeedbackWidgetProvider/index.ts
@@ -1,1 +1,0 @@
-export { FeedbackWidgetProvider, useFeedbackWidget } from "./ui/FeedbackWidgetProvider";
