+++ src/pages/FeedbackPage/index.ts
@@ -1,0 +1,1 @@
+export { FeedbackPageAsync as FeedbackPage } from "./ui/FeedbackPage/FeedbackPage.async";
