+++ src/pages/FeedbackPage/ui/FeedbackForm/FeedbackForm.module.scss
@@ -1,0 +1,34 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    width: 100%;
+    max-width: 480px;
+    margin: 60px auto;
+    padding: 32px;
+    background-color: #fff;
+    border-radius: 16px;
+    box-shadow: 0 8px 30px rgba(0, 0, 0, 0.08);
+
+    h2 {
+        margin: 0;
+        font-size: 1.5rem;
+    }
+}
+
+.description {
+    margin: 0;
+    color: var(--gray-color, #7a7a7a);
+}
+
+.submitButton {
+    padding: 12px 20px;
+    width: 100%;
+}
+
+@media (max-width: 576px) {
+    .wrapper {
+        margin: 32px auto;
+        padding: 20px;
+    }
+}
