+++ src/pages/FeedbackPage/ui/FeedbackPage/FeedbackPage.async.ts
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const FeedbackPageAsync = lazy(() => import("./FeedbackPage"));
