+++ src/pages/FeedbackPage/ui/FeedbackPage/FeedbackPage.module.scss
@@ -1,0 +1,11 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    padding-bottom: 100px;
+}
+
+.form {
+    margin-left: 20px;
+    margin-right: 20px;
+}
