+++ src/pages/FeedbackPage/ui/Header/Header.module.scss
@@ -1,0 +1,21 @@
+.wrapper {
+    width: 100%;
+    min-height: 190px;
+    background-image: linear-gradient(257deg, #EBC50C -1.78%, #00A0A8 103.74%);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.title {
+    font-size: 2.5rem;
+    text-align: center;
+    color: var(--text-secondary-1);
+    font-weight: 600;
+}
+
+@media (max-width: 576px) {
+    .title {
+        font-size: 1.625rem;
+    }
+}
