+++ src/pages/FeedbackPage/ui/Header/Header.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import cn from "classnames";
+import { useTranslation } from "react-i18next";
+import styles from "./Header.module.scss";
+
+interface HeaderProps {
+    className?: string;
+}
+
+export const Header = (props: HeaderProps) => {
+    const { className } = props;
+    const { t } = useTranslation();
+
+    return (
+        <section className={cn(className, styles.wrapper)}>
+            <h1 className={styles.title}>{t("main.welcome.header.about-project.feedback")}</h1>
+        </section>
+    );
+};
