+++ src/shared/config/routes/AppRoutes.ts
@@ -75,6 +75,7 @@ export enum AppRoutes {
     VIDEO = "video",
     PRIVACY_POLICY = "privacy_policy",
     PAYMENT_SECURITY_POLICY = "payment_security_policy",
+    FEEDBACK = "feedback",
     ACADEMY_MAIN = "academy_main",
     ACADEMY_COURSE = "academy_course",
     ACADEMY_LESSON = "academy_lesson",
