--- src/shared/ui/FeedbackModal/FeedbackModal.module.scss
@@ -1,88 +1,0 @@
-.backdrop {
-    position: fixed;
-    inset: 0;
-    z-index: 997;
-    background: transparent;
-}
-
-.panel {
-    position: fixed;
-    right: 24px;
-    bottom: 92px;
-    z-index: 999;
-
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-
-    width: 340px;
-    max-width: calc(100vw - 32px);
-    max-height: calc(100vh - 120px);
-    overflow-y: auto;
-
-    padding: 20px;
-    background-color: #fff;
-    border-radius: 16px;
-    box-shadow: 0 8px 30px rgba(0, 0, 0, 0.2);
-
-    animation: pop-in 0.15s ease-out;
-
-    h2 {
-        margin: 0;
-        padding-right: 20px;
-        font-size: 1.1rem;
-    }
-}
-
-.closeButton {
-    position: absolute;
-    top: 16px;
-    right: 16px;
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0;
-    width: 14px;
-    height: 14px;
-}
-
-.closeIcon {
-    display: flex;
-    width: 100%;
-    height: 100%;
-
-    svg path {
-        fill: #333;
-    }
-}
-
-.description {
-    margin: 0;
-    color: var(--gray-color, #7a7a7a);
-}
-
-.submitButton {
-    padding: 12px 20px;
-    width: 100%;
-}
-
-@keyframes pop-in {
-    from {
-        opacity: 0;
-        transform: translateY(8px) scale(0.98);
-    }
-
-    to {
-        opacity: 1;
-        transform: translateY(0) scale(1);
-    }
-}
-
-@media (max-width: 480px) {
-    .panel {
-        right: 16px;
-        bottom: 76px;
-        left: 16px;
-        width: auto;
-    }
-}
