--- src/widgets/FeedbackWidget/index.ts
@@ -1,1 +1,0 @@
-export { FeedbackWidgetButton } from "./ui/FeedbackWidgetButton";
