--- src/widgets/FeedbackWidget/ui/FeedbackWidgetButton.module.scss
@@ -1,85 +1,0 @@
-.wrapper {
-    position: fixed;
-    right: 24px;
-    bottom: 24px;
-    z-index: 998;
-
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-shrink: 0;
-
-    width: 56px;
-    height: 56px;
-    border: none;
-    border-radius: 50%;
-    background-color: var(--accent-color, #3dabf7);
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25);
-    cursor: pointer;
-    transition: transform 0.15s ease;
-
-    &:hover {
-        transform: scale(1.06);
-    }
-}
-
-.tooltip {
-    order: -1;
-    padding: 8px 14px;
-    border-radius: 8px;
-    background-color: #262f36;
-    color: #fff;
-    font-size: 0.85rem;
-    white-space: nowrap;
-
-    opacity: 0;
-    visibility: hidden;
-    transform: translateX(6px);
-    transition: opacity 0.15s ease, transform 0.15s ease, visibility 0.15s;
-    pointer-events: none;
-}
-
-.wrapper:hover .tooltip {
-    opacity: 1;
-    visibility: visible;
-    transform: translateX(0);
-}
-
-.icon {
-    width: 26px;
-    height: 26px;
-    display: flex;
-
-    svg path {
-        stroke: #fff;
-    }
-}
-
-@media (max-width: 768px) {
-    .wrapper {
-        right: 16px;
-        bottom: 16px;
-    }
-
-    .button {
-        width: 48px;
-        height: 48px;
-    }
-
-    .icon {
-        width: 22px;
-        height: 22px;
-    }
-
-    // Hover has no clear meaning on touch devices — hide the tooltip there
-    // rather than risk it getting stuck open after a tap.
-    .tooltip {
-        display: none;
-    }
-}
